requirements.txt
streamlit
yfinance
requests
feedparser
PyPDF2
transformers
sentence-transformers
faiss-cpu
plotly
pandas
